Restructuring packages (#17)
